Solved game for first rule (cells die with < 2 neighbours)
This commit adds logic to enable the first rule of the game: that cells die on each iteration if they have fewer than 2 live neighbours.

This is done through a GridIterator class which takes the grid through one iteration each time. Currently this holds a lot of logical and will need to be refactored once everything is there.

Also adds some smaller changes:
- refactored specs slightly to use a factory for Cell classes (using FactoryBot)
- adds a flag so that 'byebug' does not need to be required in each individual file